requirements.txt
streamlit
langchain_community
langchain_google_genai
python-dotenv
langchain
pytube
youtube_transcript_api